+++ apps/native-component-list/src/screens/UI/CommunitySegmentedControlScreen.tsx
@@ -1,48 +1,37 @@
-// @ts-nocheck
 import SegmentedControl, {
-  NativeSegmentedControlIOSChangeEvent,
-} from '@react-native-segmented-control/segmented-control';
+  type NativeSegmentedControlChangeEvent,
+} from '@expo/ui/community/segmented-control';
 import React, { useState } from 'react';
-import { ScrollView, StyleSheet, Text, View, NativeSyntheticEvent } from 'react-native';
+import { ScrollView, StyleSheet, Text, View } from 'react-native';
 
-// This example is a copy from https://github.com/react-native-community/segmented-control/blob/master/example
+// Adapted from https://github.com/react-native-community/segmented-control/blob/master/example
 
 export default function SegmentedControlScreen() {
   const [values] = useState(['One', 'Two', 'Three']);
   const [value, setValue] = useState('Unselected');
   const [selectedIndex, setIndex] = useState<number | undefined>(undefined);
 
-  const _onChange = (event: NativeSyntheticEvent<NativeSegmentedControlIOSChangeEvent>) => {
+  const _onChange = (event: NativeSegmentedControlChangeEvent) => {
     setIndex(event.nativeEvent.selectedSegmentIndex);
   };
 
-  const _onValueChange = (value: string) => {
-    setValue(value);
+  const _onValueChange = (val: string) => {
+    setValue(val);
   };
 
   return (
-    <ScrollView contentContainerStyle={styles.container}>
+    <ScrollView>
       <Text style={styles.text}>
         Note: Only the last control on this screen is expected to change state
       </Text>
 
       <View style={styles.segmentContainer}>
-        <Text style={styles.text}>Segmented controls can have values and images</Text>
-        <SegmentedControl values={['One', 'Two', require('../../assets/images/user.png')]} />
+        <Text style={styles.text}>Segmented controls can have values</Text>
+        <SegmentedControl values={['One', 'Two', 'Three']} />
       </View>
 
-      <View style={styles.segmentSection}>
-        <SegmentedControl
-          values={[
-            'One',
-            'Two',
-            'Three',
-            // It seems images higher than 18pt render stretched
-            require('../../assets/images/react-native.png'),
-            'Four',
-            'Five',
-          ]}
-        />
+      <View style={[styles.segmentSection, { backgroundColor: 'black' }]}>
+        <SegmentedControl values={['One', 'Two', 'Three', 'Four', 'Five']} appearance="dark" />
       </View>
 
       <View style={styles.segmentSection}>
@@ -51,35 +40,20 @@ export default function SegmentedControlScreen() {
       </View>
 
       <View style={styles.segmentSection}>
-        <Text style={styles.text}>Segmented controls can be momentary</Text>
-        <SegmentedControl values={['One', 'Two']} momentary />
-      </View>
-
-      <View style={styles.segmentSection}>
         <Text style={styles.text}>Segmented controls can be disabled</Text>
         <SegmentedControl enabled={false} values={['One', 'Two']} selectedIndex={1} />
       </View>
 
       <View style={styles.segmentContainer}>
-        <Text style={styles.text}>Custom colors can be provided</Text>
+        <Text style={styles.text}>Custom tint color can be provided</Text>
         <SegmentedControl
           tintColor="#ff0000"
           values={['One', 'Two', 'Three', 'Four']}
           selectedIndex={0}
-          backgroundColor="#0000ff"
-          activeTextColor="white"
         />
       </View>
       <View style={styles.segmentContainer}>
-        <SegmentedControl
-          tintColor="#00ff00"
-          values={['One', 'Two', 'Three']}
-          selectedIndex={1}
-          activeTextColor="black"
-        />
-      </View>
-      <View style={styles.segmentSection}>
-        <SegmentedControl textColor="#ff00ff" values={['One', 'Two']} selectedIndex={1} />
+        <SegmentedControl tintColor="#00ff00" values={['One', 'Two', 'Three']} selectedIndex={1} />
       </View>
 
       <View>
@@ -92,7 +66,7 @@ export default function SegmentedControlScreen() {
             onValueChange={_onValueChange}
           />
         </View>
-        <Text style={[styles.text]}>
+        <Text style={styles.text}>
           Value: {value} Index: {selectedIndex}
         </Text>
       </View>
@@ -112,8 +86,5 @@ const styles = StyleSheet.create({
   },
   segmentSection: {
     marginBottom: 25,
-  },
-  container: {
-    paddingTop: 80,
   },
 });